# admission_module/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Admission and Enrollment API",
    "description": "Backend Module for the College ERP System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/document/verify": {
      "post": {
        "tags": [
          "Admin Review"
        ],
        "summary": "Verify Application Document",
        "description": "Admin verifies or rejects a specific document uploaded by the applicant.",
        "operationId": "verify_application_document_api_v1_admin_document_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentVerificationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentVerificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload Document",
        "description": "Step 1 (Documents): Applicant uploads physical files like Aadhar or Marksheet.\nConstraints: Max 1MB, Format: jpg, jpeg, png.",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/verify": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Verify Uploaded Document",
        "description": "Step 2 (Documents): Admin verifies the uploaded document.",
        "operationId": "verify_uploaded_document_api_v1_documents_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentVerificationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentVerificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/application/{application_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Uploaded Documents",
        "description": "Allows the applicant to see all their uploaded documents before final submission.",
        "operationId": "get_uploaded_documents_api_v1_documents_application__application_id__get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response Get Uploaded Documents Api V1 Documents Application  Application Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "application_id": {
            "type": "integer",
            "title": "Application Id"
          },
          "uploader_id": {
            "type": "integer",
            "title": "Uploader Id"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "document_name": {
            "type": "string",
            "title": "Document Name"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "application_id",
          "uploader_id",
          "document_type",
          "document_name",
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
...
      "DocumentOut": {
        "properties": {
          "doc_id": {
            "type": "integer",
            "title": "Doc Id"
          },
          "application_id": {
            "type": "integer",
            "title": "Application Id"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "document_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "application_id",
          "document_type",
          "status",
          "file_path",
          "file_type",
          "file_size"
        ],
        "title": "DocumentOut"
      },
      "DocumentVerificationCreate": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "admin_id": {
            "type": "integer",
            "title": "Admin Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "admin_id",
          "status"
        ],
        "title": "DocumentVerificationCreate"
      },
      "DocumentVerificationOut": {
        "properties": {
          "verification_id": {
            "type": "integer",
            "title": "Verification Id"
          },
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "admin_id": {
            "type": "integer",
            "title": "Admin Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "verification_id",
          "document_id",
          "admin_id",
          "status",
          "remarks"
        ],
        "title": "DocumentVerificationOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}